+++ lib/reading/config.rb
@@ -167,7 +167,7 @@ module Reading
 
       {
         comment_escaped: comment_character,
-        compact_planned_row_start: /\A\s*#{comment_character}(?<genre>[^a-z:,\|]+):\s*(?=#{formats})/,
+        compact_planned_row_start: /\A\s*#{comment_character}(?:(?<genre>[^a-z:,\|]+):)?\s*(?=#{formats})/,
         compact_planned_item: /\A(?<format_emoji>(?:#{formats}))(?<author_title>[^@]+)(?<sources>@.+)?\z/,
         formats: formats,
         formats_split: /\s*(?:,|--)?\s*(?=#{formats})/,